Add session validation method in IOCP
A method to validate session ID was introduced in the IOCP classes. This function, "isValidSession", checks if a session with a given ID exists in the system. Additionally, the "FindSession" function signature was modified to allow an optional content parameter, and the visibility of "GetSessionIndex" method has been changed to static.